Add files via upload
I was learning JAVA, so i'm far away from this, but today i opened again the VS CODE and without experience or development skills i was just playing with this AI code writter.
And i think it was a good project already, need a lot o changes and implementations, but is a good start.
Maybe when i learn how to build something from scratch i could make more professional apps or websites, but now my focus was on learning Java and becoming a Fullstack developer. Thanks for you time!